+++ index.html
@@ -50,6 +50,6 @@
 
      
 
-      <script src="./src/index.js" type="module"></script>
+      <script src="./src/NotesManipulation.js" type="module"></script>
    </body>
 </html>
