+++ src/data.js
@@ -1,0 +1,61 @@
+const options = { month: 'long', day: 'numeric', year: 'numeric' };
+
+let notes = [
+   {
+      name: 'Shopping list',
+      created: new Date().toLocaleDateString('en-US', options),
+      category: 'Task',
+      content: 'tomatoes, bread',
+      dates: '',
+      command: 'sdfd',
+   },
+   {
+      name: 'The theory of evolution',
+      created: new Date().toLocaleDateString('en-US', options),
+      category: 'Random Thought',
+      content: 'The evolution...',
+      dates: '',
+      command: 'sdfd',
+   },
+   {
+      name: 'New feature',
+      created: new Date().toLocaleDateString('en-US', options),
+      category: 'Idea',
+      content: 'Implement new...',
+      dates: '3/5/2021, 5/5/2021',
+      command: 'sdfd',
+   },
+   {
+      name: 'William Gaddis',
+      created: new Date().toLocaleDateString('en-US', options),
+      category: 'Task',
+      content: "Power doesn't co",
+      dates: '',
+      command: 'sdfd',
+   },
+   {
+      name: 'Books',
+      created: new Date().toLocaleDateString('en-US', options),
+      category: 'Task',
+      content: 'The Lean Startup',
+      dates: '',
+      command: 'sdfd',
+   },
+];
+
+const categories = {
+   Task: {
+      active: 3,
+      archived: 0,
+   },
+   'Random Thought': {
+      active: 1,
+      archived: 0,
+   },
+   Idea: {
+      active: 1,
+      archived: 0,
+   },
+};
+
+export { notes, categories, options };
